Rename package to py_document_chunker (#46)
This commit renames the package from `pyDocumentChunker` to `py_document_chunker` to make it PEP conformant and ready for PyPI submission.

The following changes were made:
- The source directory was renamed from `src/pyDocumentChunker` to `src/py_document_chunker`.
- The package name in `pyproject.toml` was updated to `py_document_chunker`.
- All occurrences of `pyDocumentChunker` were replaced with `py_document_chunker` in the documentation, tests, and source code.
- Linters (`black` and `ruff`) were run to ensure code quality.
- The full test suite was run to ensure that the changes did not break any existing functionality.
